CI/CD the important components
We need a secure and safe way to be building crutial parts of the
system, the automated way.

At the moment, some of these are built by developers on their machines
and pushed to the registry...

In the future, we'd like to be notarising these images, but due to lack
of shared registry installation, we cannot go there yet.

This is a step in a good direction.